Add in support for longs (for python2) and allow both addresses and address64s to be longs.

+++ volatility/renderers/basic.py
@@ -6,16 +6,16 @@ __author__ = 'mike'
 
 from volatility import renderers
 
-class Address(int):
+class Address(long):
     """Integer class to allow renderers to differentiate between addresses and numbers"""
     def __new__(cls, number):
-        return int.__new__(cls, number)
+        return long.__new__(cls, number)
 
-class Address64(int):
+class Address64(long):
     """Integer class to allow renderers to differentiate between addresses and numbers"""
 
     def __new__(cls, number):
-        return int.__new__(cls, number)
+        return long.__new__(cls, number)
 
 
 class Hex(int):
